Added new ExcelUtility.java file and added get methods using try-with-resources.
