+++ src/test/java/com/opencart/utilities/ExcelUtility.java
@@ -1,101 +1,58 @@
 package com.opencart.utilities;
 
 import org.apache.poi.ss.usermodel.CellStyle;
-import org.apache.poi.ss.usermodel.DataFormat;
 import org.apache.poi.ss.usermodel.DataFormatter;
 import org.apache.poi.xssf.usermodel.XSSFCell;
 import org.apache.poi.xssf.usermodel.XSSFRow;
 import org.apache.poi.xssf.usermodel.XSSFSheet;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 
-import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
 
 public class ExcelUtility {
 
-    public FileInputStream fi;
-    public FileOutputStream fo;
-    public XSSFWorkbook workbook;
-    public XSSFSheet sheet;
-    public XSSFRow row;
-    public XSSFCell cell;
-    public CellStyle style;
-    String path;
+    private final String path;
+    private CellStyle style;
 
-    public ExcelUtility(String path){
+    public ExcelUtility(String path) {
         this.path = path;
     }
 
     public int getRowCount(String sheetName) throws IOException {
-        fi = new FileInputStream(path);
-        workbook = new XSSFWorkbook(fi);
-        sheet = workbook.getSheet(sheetName);
-        int rowCount = sheet.getLastRowNum();
-        workbook.close();
-        fi.close();
-        return rowCount;
+        try (FileInputStream fi = new FileInputStream(path);
+             XSSFWorkbook workbook = new XSSFWorkbook(fi)) {
+
+            XSSFSheet sheet = workbook.getSheet(sheetName);
+            return sheet.getLastRowNum();
+        }
     }
 
     public int getCellCount(String sheetName, int rowNum) throws IOException {
-        fi = new FileInputStream(path);
-       workbook = new XSSFWorkbook(fi);
-        sheet = workbook.getSheet(sheetName);
-        row = sheet.getRow(rowNum);
-        int cellCount = row.getLastCellNum();
-        workbook.close();
-        fi.close();
-        return cellCount;
+        try (FileInputStream fi = new FileInputStream(path);
+             XSSFWorkbook workbook = new XSSFWorkbook(fi)) {
+
+            XSSFSheet sheet = workbook.getSheet(sheetName);
+            XSSFRow row = sheet.getRow(rowNum);
+            return row.getLastCellNum();
+
+        }
     }
 
     public String getCellData(String sheetName, int rowNum, int colNum) throws IOException {
-        fi = new FileInputStream(path);
-        workbook = new XSSFWorkbook(fi);
-        sheet = workbook.getSheet(sheetName);
-        row = sheet.getRow(rowNum);
-        cell = row.getCell(colNum);
+        try (FileInputStream fi = new FileInputStream(path);
+             XSSFWorkbook workbook = new XSSFWorkbook(fi)) {
 
-        DataFormatter formatter = new DataFormatter();
-        String data;
+            XSSFSheet sheet = workbook.getSheet(sheetName);
+            XSSFRow row = sheet.getRow(rowNum);
+            XSSFCell cell = row.getCell(colNum);
 
-        try {
-            data = formatter.formatCellValue(cell);
-        } catch (Exception e) {
-            data = "";
+            DataFormatter formatter = new DataFormatter();
+            try {
+                return formatter.formatCellValue(cell);
+            } catch (Exception e) {
+                return "";
+            }
         }
-        workbook.close();
-        fi.close();
-        return data;
     }
-
-    public void setCellData(String sheetName, int rowNum, int colNum, String data) throws IOException {
-        File excelFile = new File(path);
-        if (!excelFile.exists()){
-            workbook = new XSSFWorkbook();
-            fo = new FileOutputStream(path);
-            workbook.write(fo);
-        }
-        fi = new FileInputStream(path);
-        workbook = new XSSFWorkbook(fi);
-
-        if (workbook.getSheetIndex(sheetName) == -1){
-            workbook.createSheet(sheetName);
-        }
-        sheet = workbook.getSheet(sheetName);
-
-        if (sheet.getRow(rowNum) == null){
-            sheet.createRow(rowNum);
-        }
-        row = sheet.getRow(rowNum);
-
-        cell = row.createCell(colNum);
-        cell.setCellValue(data);
-        fo = new FileOutputStream(path);
-        workbook.write(fo);
-        workbook.close();
-        fi.close();
-        fo.close();
-    }
-
 }
